rpc: Remove rup01.sh test
rstatd is a RPC daemon which collects performance statistics from kernel
on serving machine. On a client side there is rup (remote uptime display).

rstat client and server implementation for Linux had the latest release
4.0.1 in 2005. Although it's still installable on Debian, IMHO whole
technology is not used nowadays.

As we have other RPC tests (both RPC/TI-RPC and RPC integration, i.e.
NFS), we don't need to test this particular implementation. If this
implementation is really useful, it should be moved into libtirpc
project.

Signed-off-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>