rpc: Remove rusers01.sh test
rusersd is a RPC daemon which returns information about users currently
logged in to the system. It uses rusers on client side.

Latest release of netkit-rusers 0.17 was in 2001. Although it's still
installable on Debian, IMHO whole technology is not used nowadays.

As we have other RPC tests (both RPC/TI-RPC and RPC integration, i.e.
NFS), we don't need to test this particular implementation. If this
implementation is really useful, it should be moved into libtirpc
project.

Signed-off-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>