memcg_test_3: don't send signal when parent unexpectedly exited
When the parent process exits abnormally for some reason, the
parent process of the child process will become process 1.
In this case, the child process should not send signal to
process 1.

So we compare the realy parent pid and the saved parent pid,the
child process should exit when they has mismatched.

Signed-off-by: Lingling Li <lilingling@loongson.cn>
Signed-off-by: Hongchen Zhang <zhanghongchen@loongson.cn>