ksm: fix occasional page_volatile false positives
The KSM scan keeps running while checking the KSM status in sysfs, but
during the KSM scan the the rmap_items in the stale unstable tree of
the old pass are removed from it and are later reinserted in the new
unstable tree of the current pass, leading to some page_volatile
false positives.

The fix is stop the KSM scan temporarily while the KSM status is being
read from sysfs.

For reference here's an instance of the fixed false positives:

mem.c:255: TFAIL: pages_volatile is not 0 but 1.
mem.c:255: TFAIL: pages_unshared is not 1 but 0.

Reported-by: Eirik Fuller <efuller@redhat.com>
Co-developed-by: Li Wang <liwang@redhat.com>
Tested-by: Li Wang <liwang@redhat.com>
Tested-by: Eirik Fuller <efuller@redhat.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>