Apply runtime consistently for .tcnt != 0
This changes the test library so that runtime is applied consistently
for the tests that set .tcnt. Previously all iterations of the run()
function did run under a single timeout. This is problematic, at least,
for fuzzy sync tests that set .tcnt where the test execution window
shortens with each call of the run() function. Also this exact problem
was there even before runtime has been introduced, each subsequent call
of run() function did run with shortened execution window, we just
haven't noticed.

The whole point of runtime is to make single test instance runtime
predictable, where single instance is single execution of the run()
or run_all() function.

So this patch changes the code to call heartbeat(), which resets the
timers, before each call of run() or run_all() functions. With that
the .runtime applies over the single call of run() function as well.

Acked-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>