syscalls/timer_settime03: Scale interval with clock precision
What the test does is to:

- set initial expiration in the past
- set very small interval value
- expect the timer to overrun immediatelly many times
  to trigger timer overrun counter overflow

However the test has harcoded expectation that the kernel timers have
1ns resolution. And while that is true for many modern hardware high
resolution timers are generally not always present.

The test tried to cope with that by adding kernel requirement for
CONFIG_HIGH_RES_TIMERS=y however that does not necessarily mean that the
high resolution hardware is present or that the drivers are loaded.
This only means that the support has been compiled in the kernel.

So instead of disabling the test when kernel timers have lower precision
we scale the timer interval so that the inverval length divided by the
timer precision is constant i.e. handler_delay.

Fixes #925

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>