syscalls/fork13: Rewrite to new API + add .max_runtime
This fixes a problem we had for years, the test can run for more than
half an hour on slow systems and it used to timeout quite often.

This commit introduces a max_runtime with a 10 minute limit which is
more than enough for the test to complete on a modern hardware but at
the same time it limits the runtime to a sensible value on older
hardware and embedded.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>