read_all: Add worker timeout and rewrite scheduling
Kill and restart workers that take too long to read a file. The
default being 10% of max_runtime. A custom time can be set with the
new -t option.

This is to prevent a worker from blocking forever in a read. Currently
when this happens the whole test times out and any remaining files in
the worker's queue are not tested.

As a side effect we can now also set the timeout very low to cause
partial reads. However setting it to less than the time it takes to
start (fork) a new worker is treated as an error. Forking takes much
longer than most reads.

A number of other possible issues were fixed as well as a side effect
of changing the scheduling:

+ The worker queues are now filled in a
  "round robin" way. Before this only happened when -r was large
  enough.
+ When testing is finished the main process waits on the child procs before
  destroying the semaphores and worker queues.
+ max_runtime is set to 100 secs. This is so that the worker timeout
  is a round number.
+ Files which don't implement O_NONBLOCK and may block, no longer need
  to be avoided. Even if they refuse to die immediately;
  although this may result in TBROK due to zombie processes.

Note that with a worker timeout of 1s, 2 to 3 files will usually timeout on
my workstation. With 2s, none will. In any case testing completes in
under 3s for proc, sys or dev.

This is much faster than many other machines. It's quite likely the
worker timeout and max_runtime need to be increased on small and very
large machines. This can be done manually by setting LTP_RUNTIME_MUL.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>

Cc: Joerg Vehlow <lkml@jv-coder.de>
Cc: Jan Stancek <jstancek@redhat.com>