read_all: Allow /sys/power/wakeup_count
This was filtered because it could block at certain times. However we
will now kill the worker and continue.

This could still result in a TBROK test failure if the process can't
be killed (quick enough). Possibly this can be considered a kernel bug
because reading this file should not make an unprivileged process
unkillable. Or else the tester needs to increase LTP_RUNTIME_MUL.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Li Wang <liwang@redhat.com>