openposix: Setup autoconf and fix installation layout
This allows installation of the tests to a different directory
than directly under $prefix.

Before the layout was:
$prefix/bin/{run-*,t0,run-tests.h,Makefile}
$prefix/conformance
$prefix/functional
$prefix/stress

with prefix being /opt/openposix_testsuite on linux and
/usr/local/openposix_testsuite on other systems
OR the value of the env-var $prefix.

With this change, the prefix defaults to /opt/openposix_testsuite
and can be changed the usual way using configure (./configure ---prefix=foo)
Additionally the path of the tests below $prefix can be changed, using
configure --with-open-posix-testdir=<foo>. This allows clean installation as
part of ltp:

$prefix/bin/{run-all-posix-option-group-tests.sh,run-posix-option-group-test.sh}
$prefix/$testdir/bin/{run-tests.sh,t0}
$prefix/$testdir/conformance
$prefix/$testdir/functional
$prefix/$testdir/stress

Only the two directly callable shell scripts are kept under $prefix/bin,
the two other executables are moved to testdir, because they are only
used by the implementation of the test execution scripts (run.sh)
$prefix/bin/Makefile is not installed anymore, because it is not required.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Acked-by: Richard Palethorpe <rpalethorpe@suse.com>
Signed-off-by: Joerg Vehlow <joerg.vehlow@aox.de>