syscalls/fanotify10: Test FAN_MARK_IGNORE with FAN_EVENT_ON_CHILD
Verify correct behavior with FAN_MARK_IGNORE when FAN_EVENT_ON_CHILD
flag is set in ignore mask and not in mask and when the flag is set in
mask and not in ignore mask.

We do this for 6 test cases where ignore mark is on parent dir, but not
for the one test case where both mark and ignore mask are on parent dir.

Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>