syscalls/fanotify10: Test FAN_MARK_IGNORE with FAN_ONDIR
Verify correct behavior with FAN_MARK_IGNORE when the FAN_ONDIR flag
is set in ignore mask and not in mask and when the flag is set in
mask and not in ignore mask.

These test cases are skipped for the FAN_MARK_IGNORED_MASK variant
where ignore mask is always applied to directories regardless of
FAN_ONDIR flag in mask or ignore mask.

Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>