open04: add EMFILE check
[pre-release testing fix]

Test in automation easily get EMFILE error before reaching the fds_limit,
but hard to reproduce it again manually. The possible reason is that some
shared fd being opened in the parent shell and occupying the fd numbers
which inherited by test then results in open failed with EMFILE early.

This patch adds back of the EMFILE check in the open() loops, to flexible
test fd limitation.

  open04.c:36: TBROK: open(open04.1020,66,0777) failed: EMFILE (24)
  open04.c:53: TWARN: close(0) failed: EBADF (9)

Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Acked-by: Petr Vorel <pvorel@suse.cz>