sctp/connectx: remove invalid TEST3
TEST3 aims to check that the call returns EINVAL with invalid address
and attempts to do that by allocating the structure short of one byte.

But this does not work at all, since in case of AF_INET the last eight
bytes are padding anyways. It's probably impossible to create invalid
struct sockaddr_in (unlike invalid sa_family which is done in next
test).

Closes: https://github.com/linux-test-project/ltp/pull/975

Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Rong Tao <rongtao@cestc.cn>