build: harden ci.yml permissions
Add explicit permissions section to workflows. This is a security best
practice because by default workflows run with extended set of
permissions (except from on: pull_request from external forks).

By specifying any permission explicitly all others are set to none. By
using the principle of least privilege the damage a compromised workflow
can do (because of an injection or compromised third party tool or
action) is restricted.

It is recommended to have most strict permissions on the top level and
grant write permissions on job level case by case.

Closes: https://github.com/linux-test-project/ltp/pull/978
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Alex <aleksandrosansan@gmail.com>