tst_test.sh: Normalize the locale
network/tcp_cmds/tracepath/tracepath01.sh fails with LANG=ja_JP.UTF-8
because it parses localized output. The below is an example of such
output:
$ tracepath localhost
 1?: [LOCALHOST]                        0.040ミリ秒 pmtu 65536
 1:  localhost                                             0.274ミリ秒 到達しました
 1:  localhost                                             0.261ミリ秒 到達しました
     要約: pmtu 65536 ホップ数 1 戻りホップ数 1

It is necessary to normalize the locale to avoid such a problem.
There are some tests do the normalization, but that is not
comprehensive. Add code to normalize the locale to tst_test.sh so
that it can cover more tests.

The added code does the normalization by setting LC_ALL, which
takes precedence to the other locale-related environment variables
and does not require that "locale" command exists.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>
[ rpalethorpe: Set locale before setup ]