setfsuid02: using -1 as invalid fsuid for setfsuid()
a uid which does not have an entry in the /etc/passwd file is not really
an invalid fsuid for setfsuid(), so changing the test to use -1 as an
invalid fsuid.

And second setfsuid(-1) call is to verify that preceding call has
actually failed and there is no change in the fsuid.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Signed-off-by: Avinesh Kumar <akumar@suse.de>
[ pvorel: cast with UID_T and long ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>