syscalls/prctl02: Fix PR_CAP_AMBIENT testcases
Previously, the PR_CAP_AMBIENT testcase that expected EINVAL on nonzero
unused argument (i.e. arg3) was erring because it had an invalid arg2.
In this case, prctl still returns EINVAL, but the testcase description
doesn't match the cause. Fix the description of the testcase and add a
test that properly checks that prctl returns EINVAL on nonzero unused
argument.

Reviewed-by: Yang Xu <xuyang2018.jy@fujitsu.com>
Signed-off-by: Tudor Cretu <tudor.cretu@arm.com>