Fix aio-stress stucking with multiple threads
CPUs that increment threads_starting and threads_ending might have not
flushed the operation into cache. This causes the test to stuck when
variables have not been increased correctly. To avoid this issue we
use pthread_barrier instead of flags.
Removed also no_stonewall flag that is useless for our testing.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>