security/dirtyc0w_shmem: Fix test with host pagesize != 4k
The test can currently fail if the host page size is not 4k, because we
forgot to replace a hard-coded 4k by the actual page size. So if we end
up having to trigger a UFFDIO_WAKE, that ioctl will fail because it's
not aligned to the host page size.

Reported-by: Juri Lelli <juri.lelli@redhat.com>
Reported-by: Bruno Goncalves <bgoncalv@redhat.com>
Tested-by: Chunyu Hu <chuhu@redhat.com>
Cc: Petr Vorel <pvorel@suse.cz>
Cc: Martin Doucha <mdoucha@suse.cz>
Signed-off-by: David Hildenbrand <david@redhat.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>