cpuset/cpuset_memory_pressure_test: Fix memory calculate
Currently, free has two output formats,as follows

free -m
        total        used        free      shared  buff/cache   available
Mem:   128135        3857      120633         158        3644      123219
Swap:    8191          82        8109

free -m
        total       used       free     shared    buffers     cached
Mem:   419694       9464     410230        234        435       6005
-/+ buffers/cache:       3022     416671
Swap:    2053          0       2053

For py_mem,we need to avoid the error of adding the available item so that
py_mem is wrong. So we use the 'MemAvailable' item form '/proc/meminfo'.

For sw_mem,when swap is not inthe fourth line of the free result,it will
be NULL. So we use the 'SwapFree' item form '/proc/meminfo'.

Signed-off-by: Hao Ge <gehao@kylinos.cn>
Signed-off-by: Li Wang <liwang@redhat.com>
Reviewed-by: Li Wang <liwang@redhat.com>